requirements.txt
wheel
pandas
dash_bootstrap_components
sklearn
matplotlib
wordcloud
sklearn